requirements.txt
# SEC Filing Analyzer - Main Dependencies
# Install with: pip install -r requirements.txt
# For reproducible builds, use: pip install -r requirements.txt --no-deps

# ===========================
# SEC Filing Parsing (PINNED for reproducibility)
# ===========================
sec-parser==0.54.0
sec-downloader>=0.10.0

# ===========================
# Core NLP and ML
# ===========================
transformers>=4.35.0
torch>=2.0.0

# ===========================
# Data Handling
# ===========================
pandas>=2.0.0
numpy>=1.24.0

# ===========================
# Visualization and UI
# ===========================
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ===========================
# Text Processing
# ===========================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ===========================
# Configuration
# ===========================
PyYAML>=6.0
python-dotenv>=1.0.0

# ===========================
# Machine Learning
# ===========================
scikit-learn>=1.3.0

# ===========================
# Data Validation
# ===========================
pydantic>=2.0.0

# ===========================
# Development Dependencies (Optional)
# Install with: pip install -r requirements-dev.txt
# ===========================
